requirements.txt
fastapi == 0.115.9
uvicorn == 0.34.0
pdfplumber == 0.11.5
python-docx ==  1.1.2
google-generativeai == 0.8.4
python-multipart == 0.0.20
SQLAlchemy == 2.0.39
python-dotenv == 1.0.1
passlib == 1.7.4
pymysql == 1.1.1
bcrypt == 4.3.0
python-jose[cryptography]